+++ Assets/Prefabs/KillZones/ExplodingHealthPack1.prefab
@@ -13,7 +13,6 @@ GameObject:
   - component: {fileID: 23677006998091516}
   - component: {fileID: 64682305827614542}
   - component: {fileID: 5033351025091367117}
-  - component: {fileID: 530119411053940986}
   - component: {fileID: 3478074548994490098}
   - component: {fileID: 3126873764573027470}
   m_Layer: 0
@@ -37,6 +36,7 @@ Transform:
   m_Children:
   - {fileID: 1605662104879495672}
   - {fileID: 5449534496804672772}
+  - {fileID: 7579885129624954161}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -119,22 +119,7 @@ MonoBehaviour:
   explosionStrength: 15
   persist: 0
   ExplosionPrefab: {fileID: 1290715649332758, guid: 587eaecaa621be74486f351c78b4f0a8, type: 3}
-  ExampleCamera: {fileID: 0}
   dontRespawn: 0
---- !u!114 &530119411053940986
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1823087282522076}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 420857d4b1a36dc45869c2b7fb5a35a9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 1
-  max: Infinity
 --- !u!114 &3478074548994490098
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -148,9 +133,14 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   destroy: 0
+  consequenceObject: {fileID: 0}
+  findConsequencesInChildren: 0
   CustomTag: Player
+  acceptTagInParent: 0
   RequireItem: 
-  onStay: 0
+  reEngageOnStay: 1
+  reverseOnExit: 0
+  requireComponentsEnabled: []
 --- !u!114 &3126873764573027470
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -258,6 +248,70 @@ Light:
   m_UseViewFrustumForShadowCasterCull: 1
   m_ShadowRadius: 0
   m_ShadowAngle: 0
+--- !u!1 &7931513026162827296
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 7579885129624954161}
+  - component: {fileID: 1905119323144723785}
+  - component: {fileID: 5602831993766757111}
+  m_Layer: 0
+  m_Name: Spinner
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &7579885129624954161
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7931513026162827296}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 4856636585516382}
+  m_RootOrder: 2
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1905119323144723785
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7931513026162827296}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 420857d4b1a36dc45869c2b7fb5a35a9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  toSpin: {fileID: 1823087282522076}
+  speed: 1
+  max: Infinity
+  axis: {x: 0, y: 1, z: 0}
+--- !u!114 &5602831993766757111
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7931513026162827296}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 16b56fc371fbf8c4d80e0245173696d3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  destroy: 0
+  consequenceObject: {fileID: 0}
+  findConsequencesInChildren: 0
 --- !u!1001 &5454419465291266078
 PrefabInstance:
   m_ObjectHideFlags: 0
