+++ Assets/Prefabs/KillZones/ExplodingHealthPack2.prefab
@@ -110,7 +110,6 @@ GameObject:
   - component: {fileID: 7145503052823899730}
   - component: {fileID: 7176725549629584012}
   - component: {fileID: 3814481175828398403}
-  - component: {fileID: 3909330690062971694}
   - component: {fileID: 838385736966900192}
   - component: {fileID: 4108349191606718625}
   m_Layer: 0
@@ -134,6 +133,7 @@ Transform:
   m_Children:
   - {fileID: 334280509438265877}
   - {fileID: 8165980188421678365}
+  - {fileID: 2954167819921207503}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: -88.057, z: 0}
@@ -216,22 +216,7 @@ MonoBehaviour:
   explosionStrength: 30
   persist: 1
   ExplosionPrefab: {fileID: 1290715649332758, guid: 587eaecaa621be74486f351c78b4f0a8, type: 3}
-  ExampleCamera: {fileID: 0}
   dontRespawn: 0
---- !u!114 &3909330690062971694
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7167640292495224808}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 420857d4b1a36dc45869c2b7fb5a35a9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 1
-  max: 0
 --- !u!114 &838385736966900192
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -245,9 +230,14 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   destroy: 0
+  consequenceObject: {fileID: 0}
+  findConsequencesInChildren: 0
   CustomTag: Player
+  acceptTagInParent: 0
   RequireItem: 
-  onStay: 0
+  reEngageOnStay: 1
+  reverseOnExit: 0
+  requireComponentsEnabled: []
 --- !u!114 &4108349191606718625
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -355,3 +345,67 @@ Light:
   m_UseViewFrustumForShadowCasterCull: 1
   m_ShadowRadius: 0
   m_ShadowAngle: 0
+--- !u!1 &9049281348078571494
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 2954167819921207503}
+  - component: {fileID: 1900386636230711672}
+  - component: {fileID: 7173594349248846467}
+  m_Layer: 0
+  m_Name: Spinner
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &2954167819921207503
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9049281348078571494}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 7166296060801231918}
+  m_RootOrder: 2
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1900386636230711672
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9049281348078571494}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 420857d4b1a36dc45869c2b7fb5a35a9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  toSpin: {fileID: 7167640292495224808}
+  speed: 1
+  max: 0
+  axis: {x: 0, y: 1, z: 0}
+--- !u!114 &7173594349248846467
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9049281348078571494}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 16b56fc371fbf8c4d80e0245173696d3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  destroy: 0
+  consequenceObject: {fileID: 0}
+  findConsequencesInChildren: 0
