+++ Assets/Prefabs/RespawnLocations/ExplodingRespawnToilet Variant.prefab
@@ -1,5 +1,69 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!1 &2215223004512135764
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 2344457459336634614}
+  - component: {fileID: 2940702272464379541}
+  - component: {fileID: 8956250333870198714}
+  m_Layer: 0
+  m_Name: Spinner
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &2344457459336634614
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2215223004512135764}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 3542336818819848644}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2940702272464379541
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2215223004512135764}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 16b56fc371fbf8c4d80e0245173696d3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  destroy: 0
+  consequenceObject: {fileID: 0}
+  findConsequencesInChildren: 0
+--- !u!114 &8956250333870198714
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2215223004512135764}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 420857d4b1a36dc45869c2b7fb5a35a9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  toSpin: {fileID: 3547769243403531886}
+  speed: 1
+  max: Infinity
+  axis: {x: 0, y: 1, z: 0}
 --- !u!1001 &5861635624630967896
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -65,6 +129,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 8190fdede254466458d43ba113c18ac1, type: 3}
+--- !u!4 &3542336818819848644 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 6949157694157359004, guid: 8190fdede254466458d43ba113c18ac1, type: 3}
+  m_PrefabInstance: {fileID: 5861635624630967896}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1 &3547769243403531886 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 6945821519699830838, guid: 8190fdede254466458d43ba113c18ac1, type: 3}
@@ -83,18 +152,3 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   LocationTracksThisObject: 0
---- !u!114 &7849318225159427395
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3547769243403531886}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 420857d4b1a36dc45869c2b7fb5a35a9, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 1
-  max: Infinity
-  axis: {x: 0, y: 1, z: 0}
