--- Assets/Scripts/Consequences/RotateConsequence.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ad5d5c28adfdf945af7baf1a7403546
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
