--- Assets/Scripts/TimedEffects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3a077d5f35c417449b7103ddc9164ff3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
